ContractSectionView.jl

BitemporalReactive.ContractSectionView.ContractsModel — Type

ContractsModel

reactive data view model (Stipple/Quasar/Vue) synched between Julia Server and Browser

BitemporalReactive.ContractSectionView.contract — Method

contract contract version tab

BitemporalReactive.ContractSectionView.contract_list — Method

contract_list Display the contracts selection tab

BitemporalReactive.ContractSectionView.contract_partners — Method

contract_partners Partners, at least the policy holder, but also premium payer and beneficiaries, if different from policy holder.

BitemporalReactive.ContractSectionView.page_content — Method

page_content(model) provides the tab structure slots

BitemporalReactive.ContractSectionView.product — Method

selectedproductpart_idx Display the product tab

BitemporalReactive.ContractSectionView.product_items — Method

product_items Displays product items, technically insurance (sub) contracts of which the contract proper can contain several, for instance to represent dynamic increases of the insurance sum, which can be be revoked individually.

BitemporalReactive.ContractSectionView.tariff_item_partners — Method

tariffitempartners Display the partners referenced by a tarif item, insured person(s) typically

BitemporalReactive.ContractSectionView.tariff_items — Method

tariff_items Displays tariff parameters and partners referenced

BitemporalReactive.ContractSectionView.ui — Method

ui(model) entry point of the vue app